move database users endpoint to db.js
