+++ src/userbase-server/server.js
@@ -652,7 +652,7 @@ async function start(express, app, userbaseConfig = {}) {
     v1Admin.get('/users/:userId', admin.authenticateAccessToken, userController.adminGetUserController)
     v1Admin.get('/apps/:appId', admin.authenticateAccessToken, appController.getAppController)
     v1Admin.get('/apps/:appId/users', admin.authenticateAccessToken, appController.listUsersWithPagination)
-    v1Admin.get('/databases/:databaseId/users', admin.authenticateAccessToken, appController.listUsersForDatabaseWithPagination)
+    v1Admin.get('/databases/:databaseId/users', admin.authenticateAccessToken, db.listUsersForDatabaseWithPagination)
     v1Admin.get('/apps', admin.authenticateAccessToken, appController.listAppsWithPagination)
     v1Admin.get('/auth-tokens/:authToken', admin.authenticateAccessToken, userController.verifyAuthToken)
